+++ variables.tf
@@ -64,6 +64,12 @@ variable "s3_bucket_acl" {
   default     = "private"
 }
 
+variable "s3_shared_principal_list" {
+  type        = list(string)
+  description = "List of users/roles that will be granted permissions to GET, PUT, and DELETE objects from/to the provisioned S3 bucket."
+  default     = []
+}
+
 variable "s3_bucket_tags" {
   type        = map
   description = "Specify any tags that should be added to the S3 bucket being provisioned."